email-sig_builder/requirements.txt
pyyaml
mako
beautifulsoup4